Upload the main field and music file
The music file is basic for the program. It is designed by scientists who try create the most relaxing music ever, using the psychology rules for improving the performance for users. The main program tries to execute the binaural sound and music at the same time for peak performance of the process.